+++ lib/domain/dto/last_price/last_price_dto.freezed.dart
@@ -16,7 +16,7 @@ T _$identity<T>(T value) => value;
 /// @nodoc
 mixin _$LastPriceDto {
 
-@JsonKey(name: 'shopping_id') String get shoppingId;@JsonKey(name: 'product_id') String get productId;@JsonKey(name: 'sale_by') SaleBy get saleBy;@JsonKey(name: 'last_unit_price') int get lastUnitPrice;@JsonKey(name: 'created_at') DateTime? get createdAt;
+@JsonKey(name: 'shopping_id') String get shoppingId;@JsonKey(name: 'product_id') String get productId;@JsonKey(name: 'is_unit') bool get isUnit;@JsonKey(name: 'last_unit_price') int get lastUnitPrice;@JsonKey(name: 'created_at') DateTime? get createdAt;
 /// Create a copy of LastPriceDto
 /// with the given fields replaced by the non-null parameter values.
 @JsonKey(includeFromJson: false, includeToJson: false)
@@ -29,16 +29,16 @@ $LastPriceDtoCopyWith<LastPriceDto> get copyWith => _$LastPriceDtoCopyWithImpl<L
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is LastPriceDto&&(identical(other.shoppingId, shoppingId) || other.shoppingId == shoppingId)&&(identical(other.productId, productId) || other.productId == productId)&&(identical(other.saleBy, saleBy) || other.saleBy == saleBy)&&(identical(other.lastUnitPrice, lastUnitPrice) || other.lastUnitPrice == lastUnitPrice)&&(identical(other.createdAt, createdAt) || other.createdAt == createdAt));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is LastPriceDto&&(identical(other.shoppingId, shoppingId) || other.shoppingId == shoppingId)&&(identical(other.productId, productId) || other.productId == productId)&&(identical(other.isUnit, isUnit) || other.isUnit == isUnit)&&(identical(other.lastUnitPrice, lastUnitPrice) || other.lastUnitPrice == lastUnitPrice)&&(identical(other.createdAt, createdAt) || other.createdAt == createdAt));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,shoppingId,productId,saleBy,lastUnitPrice,createdAt);
+int get hashCode => Object.hash(runtimeType,shoppingId,productId,isUnit,lastUnitPrice,createdAt);
 
 @override
 String toString() {
-  return 'LastPriceDto(shoppingId: $shoppingId, productId: $productId, saleBy: $saleBy, lastUnitPrice: $lastUnitPrice, createdAt: $createdAt)';
+  return 'LastPriceDto(shoppingId: $shoppingId, productId: $productId, isUnit: $isUnit, lastUnitPrice: $lastUnitPrice, createdAt: $createdAt)';
 }
 
 
@@ -49,7 +49,7 @@ abstract mixin class $LastPriceDtoCopyWith<$Res>  {
   factory $LastPriceDtoCopyWith(LastPriceDto value, $Res Function(LastPriceDto) _then) = _$LastPriceDtoCopyWithImpl;
 @useResult
 $Res call({
-@JsonKey(name: 'shopping_id') String shoppingId,@JsonKey(name: 'product_id') String productId,@JsonKey(name: 'sale_by') SaleBy saleBy,@JsonKey(name: 'last_unit_price') int lastUnitPrice,@JsonKey(name: 'created_at') DateTime? createdAt
+@JsonKey(name: 'shopping_id') String shoppingId,@JsonKey(name: 'product_id') String productId,@JsonKey(name: 'is_unit') bool isUnit,@JsonKey(name: 'last_unit_price') int lastUnitPrice,@JsonKey(name: 'created_at') DateTime? createdAt
 });
 
 
@@ -66,12 +66,12 @@ class _$LastPriceDtoCopyWithImpl<$Res>
 
 /// Create a copy of LastPriceDto
 /// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? shoppingId = null,Object? productId = null,Object? saleBy = null,Object? lastUnitPrice = null,Object? createdAt = freezed,}) {
+@pragma('vm:prefer-inline') @override $Res call({Object? shoppingId = null,Object? productId = null,Object? isUnit = null,Object? lastUnitPrice = null,Object? createdAt = freezed,}) {
   return _then(_self.copyWith(
 shoppingId: null == shoppingId ? _self.shoppingId : shoppingId // ignore: cast_nullable_to_non_nullable
 as String,productId: null == productId ? _self.productId : productId // ignore: cast_nullable_to_non_nullable
-as String,saleBy: null == saleBy ? _self.saleBy : saleBy // ignore: cast_nullable_to_non_nullable
-as SaleBy,lastUnitPrice: null == lastUnitPrice ? _self.lastUnitPrice : lastUnitPrice // ignore: cast_nullable_to_non_nullable
+as String,isUnit: null == isUnit ? _self.isUnit : isUnit // ignore: cast_nullable_to_non_nullable
+as bool,lastUnitPrice: null == lastUnitPrice ? _self.lastUnitPrice : lastUnitPrice // ignore: cast_nullable_to_non_nullable
 as int,createdAt: freezed == createdAt ? _self.createdAt : createdAt // ignore: cast_nullable_to_non_nullable
 as DateTime?,
   ));
@@ -84,12 +84,12 @@ as DateTime?,
 @JsonSerializable()
 
 class _LastPriceDto extends LastPriceDto {
-  const _LastPriceDto({@JsonKey(name: 'shopping_id') required this.shoppingId, @JsonKey(name: 'product_id') required this.productId, @JsonKey(name: 'sale_by') required this.saleBy, @JsonKey(name: 'last_unit_price') required this.lastUnitPrice, @JsonKey(name: 'created_at') this.createdAt}): super._();
+  const _LastPriceDto({@JsonKey(name: 'shopping_id') required this.shoppingId, @JsonKey(name: 'product_id') required this.productId, @JsonKey(name: 'is_unit') this.isUnit = true, @JsonKey(name: 'last_unit_price') required this.lastUnitPrice, @JsonKey(name: 'created_at') this.createdAt}): super._();
   factory _LastPriceDto.fromJson(Map<String, dynamic> json) => _$LastPriceDtoFromJson(json);
 
 @override@JsonKey(name: 'shopping_id') final  String shoppingId;
 @override@JsonKey(name: 'product_id') final  String productId;
-@override@JsonKey(name: 'sale_by') final  SaleBy saleBy;
+@override@JsonKey(name: 'is_unit') final  bool isUnit;
 @override@JsonKey(name: 'last_unit_price') final  int lastUnitPrice;
 @override@JsonKey(name: 'created_at') final  DateTime? createdAt;
 
@@ -106,16 +106,16 @@ Map<String, dynamic> toJson() {
 
 @override
 bool operator ==(Object other) {
-  return identical(this, other) || (other.runtimeType == runtimeType&&other is _LastPriceDto&&(identical(other.shoppingId, shoppingId) || other.shoppingId == shoppingId)&&(identical(other.productId, productId) || other.productId == productId)&&(identical(other.saleBy, saleBy) || other.saleBy == saleBy)&&(identical(other.lastUnitPrice, lastUnitPrice) || other.lastUnitPrice == lastUnitPrice)&&(identical(other.createdAt, createdAt) || other.createdAt == createdAt));
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _LastPriceDto&&(identical(other.shoppingId, shoppingId) || other.shoppingId == shoppingId)&&(identical(other.productId, productId) || other.productId == productId)&&(identical(other.isUnit, isUnit) || other.isUnit == isUnit)&&(identical(other.lastUnitPrice, lastUnitPrice) || other.lastUnitPrice == lastUnitPrice)&&(identical(other.createdAt, createdAt) || other.createdAt == createdAt));
 }
 
 @JsonKey(includeFromJson: false, includeToJson: false)
 @override
-int get hashCode => Object.hash(runtimeType,shoppingId,productId,saleBy,lastUnitPrice,createdAt);
+int get hashCode => Object.hash(runtimeType,shoppingId,productId,isUnit,lastUnitPrice,createdAt);
 
 @override
 String toString() {
-  return 'LastPriceDto(shoppingId: $shoppingId, productId: $productId, saleBy: $saleBy, lastUnitPrice: $lastUnitPrice, createdAt: $createdAt)';
+  return 'LastPriceDto(shoppingId: $shoppingId, productId: $productId, isUnit: $isUnit, lastUnitPrice: $lastUnitPrice, createdAt: $createdAt)';
 }
 
 
@@ -126,7 +126,7 @@ abstract mixin class _$LastPriceDtoCopyWith<$Res> implements $LastPriceDtoCopyWi
   factory _$LastPriceDtoCopyWith(_LastPriceDto value, $Res Function(_LastPriceDto) _then) = __$LastPriceDtoCopyWithImpl;
 @override @useResult
 $Res call({
-@JsonKey(name: 'shopping_id') String shoppingId,@JsonKey(name: 'product_id') String productId,@JsonKey(name: 'sale_by') SaleBy saleBy,@JsonKey(name: 'last_unit_price') int lastUnitPrice,@JsonKey(name: 'created_at') DateTime? createdAt
+@JsonKey(name: 'shopping_id') String shoppingId,@JsonKey(name: 'product_id') String productId,@JsonKey(name: 'is_unit') bool isUnit,@JsonKey(name: 'last_unit_price') int lastUnitPrice,@JsonKey(name: 'created_at') DateTime? createdAt
 });
 
 
@@ -143,12 +143,12 @@ class __$LastPriceDtoCopyWithImpl<$Res>
 
 /// Create a copy of LastPriceDto
 /// with the given fields replaced by the non-null parameter values.
-@override @pragma('vm:prefer-inline') $Res call({Object? shoppingId = null,Object? productId = null,Object? saleBy = null,Object? lastUnitPrice = null,Object? createdAt = freezed,}) {
+@override @pragma('vm:prefer-inline') $Res call({Object? shoppingId = null,Object? productId = null,Object? isUnit = null,Object? lastUnitPrice = null,Object? createdAt = freezed,}) {
   return _then(_LastPriceDto(
 shoppingId: null == shoppingId ? _self.shoppingId : shoppingId // ignore: cast_nullable_to_non_nullable
 as String,productId: null == productId ? _self.productId : productId // ignore: cast_nullable_to_non_nullable
-as String,saleBy: null == saleBy ? _self.saleBy : saleBy // ignore: cast_nullable_to_non_nullable
-as SaleBy,lastUnitPrice: null == lastUnitPrice ? _self.lastUnitPrice : lastUnitPrice // ignore: cast_nullable_to_non_nullable
+as String,isUnit: null == isUnit ? _self.isUnit : isUnit // ignore: cast_nullable_to_non_nullable
+as bool,lastUnitPrice: null == lastUnitPrice ? _self.lastUnitPrice : lastUnitPrice // ignore: cast_nullable_to_non_nullable
 as int,createdAt: freezed == createdAt ? _self.createdAt : createdAt // ignore: cast_nullable_to_non_nullable
 as DateTime?,
   ));