+++ lib/domain/dto/last_price/last_price_dto.g.dart
@@ -10,7 +10,7 @@ _LastPriceDto _$LastPriceDtoFromJson(Map<String, dynamic> json) =>
     _LastPriceDto(
       shoppingId: json['shopping_id'] as String,
       productId: json['product_id'] as String,
-      saleBy: $enumDecode(_$SaleByEnumMap, json['sale_by']),
+      isUnit: json['is_unit'] as bool? ?? true,
       lastUnitPrice: (json['last_unit_price'] as num).toInt(),
       createdAt: json['created_at'] == null
           ? null
@@ -21,9 +21,7 @@ Map<String, dynamic> _$LastPriceDtoToJson(_LastPriceDto instance) =>
     <String, dynamic>{
       'shopping_id': instance.shoppingId,
       'product_id': instance.productId,
-      'sale_by': _$SaleByEnumMap[instance.saleBy]!,
+      'is_unit': instance.isUnit,
       'last_unit_price': instance.lastUnitPrice,
       'created_at': instance.createdAt?.toIso8601String(),
     };
-
-const _$SaleByEnumMap = {SaleBy.unit: 'unit', SaleBy.weight: 'weight'};