+++ lib/domain/models/item/item_model.g.dart
@@ -10,7 +10,7 @@ _ItemModel _$ItemModelFromJson(Map<String, dynamic> json) => _ItemModel(
   shoppingId: json['shopping_id'] as String,
   productId: json['product_id'] as String,
   name: json['name'] as String,
-  saleBy: $enumDecodeNullable(_$SaleByEnumMap, json['sale_by']) ?? SaleBy.unit,
+  isUnit: json['sale_by'] as bool? ?? true,
   unitPrince: (json['unit_price'] as num).toInt(),
   quantity: (json['quantity'] as num?)?.toInt() ?? 1,
   priceVariation: (json['price_variation'] as num?)?.toDouble() ?? 0,
@@ -24,11 +24,9 @@ Map<String, dynamic> _$ItemModelToJson(_ItemModel instance) =>
       'shopping_id': instance.shoppingId,
       'product_id': instance.productId,
       'name': instance.name,
-      'sale_by': _$SaleByEnumMap[instance.saleBy]!,
+      'sale_by': instance.isUnit,
       'unit_price': instance.unitPrince,
       'quantity': instance.quantity,
       'price_variation': instance.priceVariation,
       'created_at': instance.createdAt?.toIso8601String(),
     };
-
-const _$SaleByEnumMap = {SaleBy.unit: 'unit', SaleBy.weight: 'weight'};