+++ lib/domain/models/last_price/last_price_model.g.dart
@@ -12,7 +12,7 @@ _LastPriceModel _$LastPriceModelFromJson(Map<String, dynamic> json) =>
       shoppingId: json['shopping_id'] as String,
       productId: json['product_id'] as String,
       lastUnitPrice: (json['last_unit_price'] as num).toInt(),
-      saleBy: $enumDecode(_$SaleByEnumMap, json['sale_by']),
+      isUnit: json['is_unit'] as bool? ?? true,
       createdAt: json['created_at'] == null
           ? null
           : DateTime.parse(json['created_at'] as String),
@@ -24,8 +24,6 @@ Map<String, dynamic> _$LastPriceModelToJson(_LastPriceModel instance) =>
       'shopping_id': instance.shoppingId,
       'product_id': instance.productId,
       'last_unit_price': instance.lastUnitPrice,
-      'sale_by': _$SaleByEnumMap[instance.saleBy]!,
+      'is_unit': instance.isUnit,
       'created_at': instance.createdAt?.toIso8601String(),
     };
-
-const _$SaleByEnumMap = {SaleBy.unit: 'unit', SaleBy.weight: 'weight'};